[NCL-7650]: Change ProductMilestoneStatistics#artifactsInMilestone from Integer to Long
This way no explicit type cast to Integer is needed when using Criteria query's count()

+++ dto/src/main/java/org/jboss/pnc/dto/response/statistics/ProductMilestoneStatistics.java
@@ -41,7 +41,7 @@ public class ProductMilestoneStatistics {
     /**
      * Number of artifacts produced by builds contained in this milestone.
      */
-    Integer artifactsInMilestone;
+    Long artifactsInMilestone;
 
     /**
      * Statistics about this milestone's delivered artifacts.
